Enable C++11 and set OpenMVG as an external dependency
