+++ src/CMakeLists.txt
@@ -1,14 +1,12 @@
 
 cmake_minimum_required(VERSION 2.6.2)
-project(openMVG-UI)
-
+project(mayaMVG)
 
 #
 # Compiler settings
 #
 if(UNIX)
-	set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -fPIC")
-	set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -fPIC")
+	set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -fPIC -std=c++11")
 endif()
 
 #
@@ -20,5 +18,4 @@ set(CMAKE_MODULE_PATH "${MAINFOLDER}/cmake")
 #
 # Add Build Targets
 #
-add_subdirectory(openMVG/src EXCLUDE_FROM_ALL)
 add_subdirectory(mayaMVG)
