+++ src/mayaMVG/core/MVGCamera.hpp
@@ -1,8 +1,6 @@
 #pragma once
 
-#undef Success // needed by eigen
 #include "mayaMVG/core/MVGNodeWrapper.hpp"
-#include "openMVG/cameras/PinholeCamera.hpp"
 #include <vector>
 #include <map>
