+++ src/mayaMVG/core/MVGLineConstrainedPlaneKernel.hpp
@@ -1,7 +1,7 @@
 #pragma once
+
+#include "MVGEigen.hpp"
 #include <maya/MPoint.h>
-#include <openMVG/numeric/numeric.h>
-#include <openMVG/robust_estimation/robust_estimator_LMeds.hpp>
 
 namespace mayaMVG
 {
@@ -15,43 +15,10 @@ struct LineConstrainedPlaneKernel
         MINIMUM_SAMPLES = 1
     };
     LineConstrainedPlaneKernel(const openMVG::Mat& pt, const openMVG::Vec3& constraintP0,
-                               const openMVG::Vec3& constraintP1)
-        : _pt(pt)
-        , _constraintP0(constraintP0)
-        , _constraintP1(constraintP1)
-        // Compute the segment values (in 3d) between p1 and p0
-        , _P1P0(_constraintP1 - _constraintP0)
-    {
-    }
+                               const openMVG::Vec3& constraintP1);
     size_t NumSamples() const { return _pt.cols(); }
-    void Fit(const std::vector<size_t>& samples, std::vector<Model>* equation) const
-    {
-        assert(samples.size() >= MINIMUM_SAMPLES);
-        equation->clear();
-        openMVG::Mat sampled_xs = openMVG::ExtractColumns(_pt, samples);
-        openMVG::Vec3 p2 = sampled_xs.col(0);
-        // Compute the segment values (in 3d) between p2 and p0
-        const openMVG::Vec3 p2p0 = p2 - _constraintP0;
-        // Avoid some crashes by checking for collinearity here
-        openMVG::Vec3 dy1dy2 = _P1P0.array() / p2p0.array();
-        if((dy1dy2[0] == dy1dy2[1]) && (dy1dy2[2] == dy1dy2[1])) // Check for collinearity
-            return;
-        // Compute the plane coefficients from the 3 given points in a
-        // straightforward manner
-        // calculate the plane normal n = (p2-p1) x (p3-p1) = cross(p2-p1, p3-p1)
-        Model m;
-        m[0] = _P1P0[1] * p2p0[2] - _P1P0[2] * p2p0[1];
-        m[1] = _P1P0[2] * p2p0[0] - _P1P0[0] * p2p0[2];
-        m[2] = _P1P0[0] * p2p0[1] - _P1P0[1] * p2p0[0];
-        m[3] = 0.0;
-        m.normalize();
-        // [ n' d ] dot [ p0 ; 1 ])
-        // m[3] = -d with d the distance from the plane m, whose unit normal is n, to p0
-        //        m[3] = -1.0 * (m.head<3>().dot(p0));
-        m[3] = -1.0 * (m.head<3>().dot(_constraintP0));
-        equation->push_back(m);
-    }
-    double Error(size_t sample, const Model& model) const
+    void Fit(const std::vector<size_t>& samples, std::vector<Model>* equation) const;
+    inline double Error(size_t sample, const Model& model) const
     {
         // Calculate the distance from the point to the plane normal as the dot
         // product
